+++ StripeFinancialConnections/StripeFinancialConnections/Source/Common/HostController.swift
@@ -129,7 +129,8 @@ class HostController {
         configuration: FinancialConnectionsSheet.Configuration,
         elementsSessionContext: ElementsSessionContext?,
         publishableKey: String?,
-        stripeAccount: String?
+        stripeAccount: String?,
+        analyticsClient: FinancialConnectionsAnalyticsClient = FinancialConnectionsAnalyticsClient()
     ) {
         self.apiClient = apiClient
         self.analyticsClientV1 = analyticsClientV1
@@ -137,7 +138,7 @@ class HostController {
         self.returnURL = returnURL
         self.configuration = configuration
         self.elementsSessionContext = elementsSessionContext
-        self.analyticsClient = FinancialConnectionsAnalyticsClient()
+        self.analyticsClient = analyticsClient
         analyticsClient.setAdditionalParameters(
             publishableKey: publishableKey,
             stripeAccount: stripeAccount
@@ -173,8 +174,9 @@ extension HostController: HostViewControllerDelegate {
         _ viewController: HostViewController,
         didFetch synchronizePayload: FinancialConnectionsSynchronize
     ) {
-        delegate?.hostController(self, didReceiveEvent: FinancialConnectionsEvent(name: .open))
         self.linkAccountSessionId = synchronizePayload.manifest.id
+        analyticsClient.setAdditionalParameters(fromManifest: synchronizePayload.manifest)
+        publish(FinancialConnectionsEvent(name: .open))
 
         let flowRouter = FlowRouter(
             synchronizePayload: synchronizePayload,
@@ -203,7 +205,7 @@ extension HostController: HostViewControllerDelegate {
         _ hostViewController: HostViewController,
         didReceiveEvent event: FinancialConnectionsEvent
     ) {
-        delegate?.hostController(self, didReceiveEvent: event)
+        publish(event)
     }
 }
 
@@ -211,13 +213,16 @@ extension HostController: HostViewControllerDelegate {
 
 private extension HostController {
 
-    func continueWithWebFlow(_ manifest: FinancialConnectionsSessionManifest, prefillDetails: WebPrefillDetails? = nil) {
-        delegate?.hostController(
-            self,
-            didReceiveEvent: FinancialConnectionsEvent(
-                name: .flowLaunchedInBrowser
-            )
+    func publish(_ event: FinancialConnectionsEvent) {
+        analyticsClient.logExternalEvent(
+            event,
+            pane: FinancialConnectionsAnalyticsClient.paneFromViewController(navigationController.topViewController)
         )
+        delegate?.hostController(self, didReceiveEvent: event)
+    }
+
+    func continueWithWebFlow(_ manifest: FinancialConnectionsSessionManifest, prefillDetails: WebPrefillDetails? = nil) {
+        publish(FinancialConnectionsEvent(name: .flowLaunchedInBrowser))
 
         let accountFetcher = FinancialConnectionsAccountAPIFetcher(api: apiClient, clientSecret: clientSecret)
         let sessionFetcher = FinancialConnectionsSessionAPIFetcher(
@@ -283,7 +288,7 @@ extension HostController: FinancialConnectionsWebFlowViewControllerDelegate {
         _ webFlowViewController: UIViewController,
         didReceiveEvent event: FinancialConnectionsEvent
     ) {
-        delegate?.hostController(self, didReceiveEvent: event)
+        publish(event)
     }
 }
 
@@ -311,7 +316,7 @@ extension HostController: NativeFlowControllerDelegate {
         _ nativeFlowController: NativeFlowController,
         didReceiveEvent event: FinancialConnectionsEvent
     ) {
-        delegate?.hostController(self, didReceiveEvent: event)
+        publish(event)
     }
 
     func nativeFlowController(
@@ -331,6 +336,6 @@ extension HostController: FinancialConnectionsAnalyticsClientDelegate {
         _ analyticsClient: FinancialConnectionsAnalyticsClient,
         didReceiveEvent event: FinancialConnectionsEvent
     ) {
-        delegate?.hostController(self, didReceiveEvent: event)
+        publish(event)
     }
 }